requirements_planning.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
pathlib>=1.0.1
jsonschema>=4.0.0
python-dotenv>=1.0.0

# Document processing
unstructured>=0.10.0
unstructured[pdf,docx,pptx]>=0.10.0
layoutparser>=0.3.4
pdf2image>=1.16.3
beautifulsoup4>=4.10.0
pdfminer.six>=20220524
pypdfium2>=4.10.0

# Transformer models
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2

# Geospatial processing
geopandas>=0.10.0
shapely>=2.0.0
pyproj>=3.3.0

# LLM API access
openai>=1.3.0
anthropic>=0.5.0

# Visualization tools (optional)
matplotlib>=3.5.0
seaborn>=0.12.0 